+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ## vN.N.N
+* add way to get from image picker to image asset page.
 * Adding support for article labels.